move all screens to a new folder screens outside the component folder
