+++ .expo/packager-info.json
@@ -1,0 +1,10 @@
+{
+  "expoServerPort": 19000,
+  "expoServerNgrokUrl": null,
+  "packagerNgrokUrl": null,
+  "ngrokPid": null,
+  "packagerPort": 19001,
+  "packagerPid": 6750,
+  "webpackServerPort": null,
+  "devToolsPort": 19002
+}
