+++ .expo/settings.json
@@ -1,0 +1,9 @@
+{
+  "hostType": "lan",
+  "lanType": "ip",
+  "dev": true,
+  "minify": false,
+  "urlRandomness": "v4-bgs",
+  "https": false,
+  "scheme": null
+}
